change: rename AddOutputQueueDialog to AddOutputProfileDialog
