+++ src/Beutl/Views/Dialogs/AddOutputProfileDialog.axaml
@@ -1,0 +1,39 @@
+<ui:ContentDialog x:Class="Beutl.Views.Dialogs.AddOutputProfileDialog"
+                  xmlns="https://github.com/avaloniaui"
+                  xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+                  xmlns:converters="using:Beutl.Converters"
+                  xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+                  xmlns:icons="using:FluentIcons.FluentAvalonia"
+                  xmlns:lang="using:Beutl.Language"
+                  xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+                  xmlns:ui="using:FluentAvalonia.UI.Controls"
+                  xmlns:vm="using:Beutl.ViewModels.Dialogs"
+                  Title="{x:Static lang:Strings.AddOutputQueue}"
+                  d:DesignHeight="450"
+                  d:DesignWidth="800"
+                  CloseButtonText="{x:Static lang:Strings.Close}"
+                  PrimaryButtonText="{x:Static lang:Strings.Add}"
+                  mc:Ignorable="d">
+    <ui:ContentDialog.Resources>
+        <StaticResource x:Key="ContentDialogMaxWidth" ResourceKey="ContentDialogMinWidth" />
+    </ui:ContentDialog.Resources>
+
+    <StackPanel Spacing="8">
+        <TextBlock Text="{x:Static lang:Strings.Extensions}" />
+        <ListBox MaxHeight="235"
+                 ItemsSource="{Binding AvailableExtensions}"
+                 SelectedItem="{Binding SelectedExtension.Value}"
+                 SelectionMode="AlwaysSelected">
+            <ListBox.ItemTemplate>
+                <DataTemplate>
+                    <StackPanel Margin="0,4">
+                        <TextBlock Text="{Binding Name}" />
+                        <TextBlock FontSize="12"
+                                   Text="{Binding}"
+                                   Theme="{StaticResource LabelTextBlockStyle}" />
+                    </StackPanel>
+                </DataTemplate>
+            </ListBox.ItemTemplate>
+        </ListBox>
+    </StackPanel>
+</ui:ContentDialog>
